% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.adonis.R
\name{pairwise.adonis}
\alias{pairwise.adonis}
\title{Pairwise multilevel comparison using adonis}
\usage{
pairwise.adonis(x, factors, sim.function = "vegdist", sim.method = "bray",
  p.adjust.m = "bonferroni")
}
\arguments{
\item{x}{Data frame (the community table).}

\item{factors}{Vector (a column or vector with the levels to be compared pairwise).}

\item{sim.function}{Function used to calculate the similarity matrix,
one of 'daisy' or 'vegdist' default is 'vegdist'.}

\item{sim.method}{Similarity method from daisy or vegdist, default is 'bray'.}

\item{p.adjust.m}{The p.value correction method, one of the methods supported by p.adjust(),
default is 'bonferroni'.}
}
\value{
Table with the pairwise factors, F-values, R^2, p.value and adjusted p.value.
}
\description{
This is a wrapper function for multilevel pairwise comparison
using adonis() from package 'vegan'. The function returns adjusted p-values using p.adjust().
}
\examples{
data(iris)
pairwise.adonis(iris[,1:4],iris$Species)

#similarity euclidean from vegdist and holm correction
pairwise.adonis(x=iris[,1:4],factors=iris$Species,sim.function='vegdist',
sim.method='euclidian',p.adjust.m='holm')

#similarity manhattan from daisy and bonferroni correction
pairwise.adonis(x=iris[,1:4],factors=iris$Species,sim.function='daisy',
sim.method='manhattan',p.adjust.m='bonferroni')
}
\author{
Pedro Martinez Arbizu
}
