% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.adonis2.R
\name{pairwise.adonis2}
\alias{pairwise.adonis2}
\title{Pairwise multilevel comparison using adonis accepting strata}
\usage{
pairwise.adonis2(x, data, strata = NULL, ...)
}
\arguments{
\item{x}{Model formula. The LHS is either community matrix or dissimilarity matrix (eg. from vegdist or dist)
See adonis() for details. The RHS are factors that nmust be column names of a data.frame specified with argument data.}

\item{data}{The data frame of indipendent varibles having as column names the factors specified in formula.}

\item{strata}{String. The name of the column with factors to be used as strata.}

\item{(...)}{Any other parameter passed to adonis}
}
\value{
List. Elements are the summary returned by adonis for each unique pairwise combination of factors.
}
\description{
This is a wrapper function for multilevel pairwise comparison
using adonis() from package 'vegan'. The function accepts interaction between factors and strata.
}
\examples{
data(iris)
pairwise.adonis2(iris[,1:4]~Species,data=iris)

# For strata, extract factors into a new data.frame
fac <- data.frame(Species=iris$Species,
	Season = rep( c('winter','summer'),75))

pairwise.adonis2(dist(iris[,1:4])~Species/Season,data=fac,strata='Season')

# Note: do not define strata indexing a column e.g. strata = iris[,5]. This will raise an error.
# Use the name of the column as string as example above.

}
\author{
Pedro Martinez Arbizu
}
